requirements.txt
certifi==2021.5.30
cycler==0.10.0
gunicorn==20.1.0
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.4.2
numpy==1.20.3
pandas==1.2.4
Pillow==8.2.0
plotly==4.14.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
retrying==1.3.3
scikit-learn==0.24.2
scipy==1.6.3
six==1.16.0
threadpoolctl==2.1.0
torch==1.8.1
torchvision==0.9.1
typing-extensions==3.10.0.0
wincertstore==0.2
